+++ templates/web/static/base/public/js/index.js
@@ -1,3 +1,4 @@
+//This file is /public/js/index.js
 (() => {
   console.log('🚀 App loaded');
 })();